# im-backoffice/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  auth.SignInReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.SignInResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    required:
    - access_token
    type: object
  auth.SignOutDTO:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
  errors.Response:
    properties:
      key:
        type: string
      message:
        type: string
    type: object
  inventory.CreateInventoryRequest:
    properties:
      actual_qty:
        type: string
      inventory_category_id:
        type: string
      inventory_name:
        type: string
      inventory_rfid:
        type: string
      location_id:
        type: string
      logistic_center_id:
        type: string
      warehouse_id:
        type: string
    required:
    - actual_qty
    - inventory_category_id
    - inventory_name
    - inventory_rfid
    - location_id
    - logistic_center_id
    - warehouse_id
    type: object
  inventory.InventoryInfoDto:
    properties:
      actual_qty:
        type: string
      inventory_id:
        type: string
      inventory_name:
        type: string
      inventory_rfid:
        type: string
      location:
        $ref: '#/definitions/inventory.LocationDto'
      logistic_center:
        $ref: '#/definitions/logisticCenter.LogisticCenterRes'
      warehouse:
        $ref: '#/definitions/inventory.WarehouseDto'
    type: object
  inventory.InventoryResponse:
    properties:
      inventories:
        items:
          $ref: '#/definitions/inventory.InventoryInfoDto'
        type: array
      inventory_category:
        $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
    type: object
  inventory.LocationDto:
    properties:
      location_id:
        type: string
      location_name:
        type: string
    type: object
  inventory.UpdateInventoryRequest:
    properties:
      actual_qty:
        type: string
      inventory_category_id:
        type: string
      inventory_name:
        type: string
      location_id:
        type: string
      logistic_center_id:
        type: string
      warehouse_id:
        type: string
    required:
    - actual_qty
    - inventory_category_id
    - inventory_name
    - location_id
    - logistic_center_id
    - warehouse_id
    type: object
  inventory.WarehouseDto:
    properties:
      warehouse_id:
        type: string
      warehouse_name:
        type: string
    type: object
  inventoryCategory.InventoryCategoryDTO:
    properties:
      id:
        type: string
      inventory_category_name:
        type: string
    type: object
...
  logisticCenter.LogisticCenterRes:
    properties:
      id:
        type: string
      logistic_center_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an inventory management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory Management API
  version: "2.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshTokenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: RefreshToken refreshes the given token.
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign In Req
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.SignInReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SignInResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      summary: Sign in the user to system.
      tags:
      - auth
  /auth/sign-out:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SignOutDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Sign out the user from system.
      tags:
      - auth
  /inventories/{page}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Search text (length min=3)
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.InventoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all inventories.
      tags:
      - inventory
  /inventory:
    post:
      consumes:
      - application/json
      parameters:
      - description: Inventory
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inventory.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new inventory.
      tags:
      - inventory
...
  /inventory/{inventoryId}:
    delete:
      parameters:
      - description: Inventory ID
        in: path
        name: inventoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete inventory by id.
      tags:
      - inventory
    get:
      consumes:
      - application/json
      parameters:
      - description: Inventory ID
        in: path
        name: inventoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Gets an inventory by ID.
      tags:
      - inventory
    put:
      consumes:
      - application/json
      parameters:
      - description: Inventory DTO
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inventory.UpdateInventoryRequest'
      - description: Inventory ID
        in: path
        name: inventoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the inventory.
      tags:
      - inventory
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"